+++ README.md
@@ -4,14 +4,10 @@ React version of course-reviews
 # Install & Run
 Make sure you have [Meteor](https://www.meteor.com) installed!
 
-Uncomment lines 322-325 in `/imports/api/classes.js`
-
     npm install
     meteor npm install
     meteor --port 3000
 
-To populate your local database with course data.
+To populate your local database with course data, uncomment lines 274-277 in `/imports/api/classes.js`. Do not make changes to the code as classes are being added (once there are no additional "new class: ..." statements printed in the terminal, all classes have been added). Re-comment the above lines to prevent continuous repopulation of the database.
 
-Do not make changes to the code as classes are being added. Once there are no additional "new class: ..." statements printed in the terminal, all classes have been added. Re-comment the above lines to prevent continuous repopulation of the database.
-
-You can access the app at http://localhost:3000
+You can access the app at http://localhost:3000.
